fixed issue with low precision normals interpolation

+++ WickedEngine/shaders/envMap_skyGS_emulation.hlsl
@@ -5,14 +5,14 @@
 struct GSInput
 {
 	float4 pos : SV_POSITION;
-	min16float3 nor : NORMAL;
+	float3 nor : NORMAL;
 	uint RTIndex : RTINDEX;
 };
 
 struct GSOutput
 {
 	float4 pos : SV_POSITION;
-	min16float3 nor : NORMAL;
+	float3 nor : NORMAL;
 	uint RTIndex : SV_RenderTargetArrayIndex;
 };
