+++ WickedEngine/shaders/objectPS_voxelizer.hlsl
@@ -45,7 +45,7 @@ struct PSInput
 	float4 pos : SV_POSITION;
 	centroid float4 uvsets : UVSETS;
 	centroid min16float4 color : COLOR;
-	centroid min16float3 N : NORMAL;
+	centroid float3 N : NORMAL;
 	centroid float3 P : POSITION3D;
 
 #ifdef VOXELIZATION_CONSERVATIVE_RASTERIZATION_ENABLED
